% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sh.R
\name{sh}
\alias{sh}
\title{Launch a linux bash shell}
\usage{
sh(
  command = NULL,
  current_dir = getwd(),
  complete_path = FALSE,
  intern = FALSE
)
}
\arguments{
\item{command}{call system and returns}

\item{current_dir}{dir of execution}

\item{complete_path}{prompts the absolute path of current directory}

\item{intern}{returns shell output (see system)}
}
\value{
nothing or shell ouput
}
\description{
Launch a linux bash shell
}
\examples{
sh()
}
